survey_platform/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
pillow==10.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.8
drf_social_oauth2==2.1.3
drf-yasg==1.21.7
django-filter==23.3
flake8==6.1.0
djangorestframework-simplejwt==5.3.0
coverage==7.3.1